fix: Fix spawned enemies killing each other
During the `The Beasts Inside` event, in very rare occasions, the dogs spawn next to each other. For each enemy that is spawned, I assign them a new EnemyType. Inside of `MouthDogAI.OnCollideWithEnemy` (which is called when a dog touches another enemy), it checks to see if the `EnemyType`'s are equal (which should always be the case), but because I created new `EnemyType`'s for each enemy spawn, this fails and kills both dogs in the process, because it thinks another enemy is killing this dog.